+++ frontend/web/components/base/forms/Tabs.js
@@ -95,9 +95,9 @@ const Tabs = class extends React.Component {
             return (
               <div
                 key={`content${i}`}
-                className={`tab-item ${isSelected ? ' tab-active' : ''} ${
+                className={`${isSelected ? ' tab-active' : ''} ${
                   this.props.isRoles && 'p-0'
-                }`}
+                } ${child.props.className || 'tab-item'}`}
               >
                 {child}
               </div>
